--- src/main/java/com/kh/hobbycloud/entity/notice/NoticeVO.java
@@ -1,18 +1,0 @@
-package com.kh.hobbycloud.entity.notice;
-
-import java.sql.Date;
-
-import lombok.Data;
-
-@Data
-public class NoticeVO {
-    private int noticeIdx;
-	private String noticeName;
-	private String noticeDetail;
-	private int memberIdx;
-	private String memberNick;
-	private Date noticeRegistered;
-	private int noticeViews;
-	private int noticeReplies;
-
-}
