+++ src/main/java/com/kh/hobbycloud/repository/notice/NoticeDao.java
@@ -3,7 +3,7 @@ package com.kh.hobbycloud.repository.notice;
 import java.util.List;
 
 import com.kh.hobbycloud.entity.notice.NoticeDto;
-import com.kh.hobbycloud.entity.notice.NoticeVO;
+import com.kh.hobbycloud.vo.notice.NoticeVO;
 
 public interface NoticeDao {
 	List<NoticeVO>list();
