+++ src/main/java/com/kh/hobbycloud/vo/notice/NoticeVO.java
@@ -1,18 +1,20 @@
-package com.kh.hobbycloud.vo;
+package com.kh.hobbycloud.vo.notice;
 
 import java.sql.Date;
 
 import org.springframework.web.multipart.MultipartFile;
 
 import lombok.Data;
+
 @Data
-public class NoticeFileVO {
-	private int noticeIdx;
-	private int memberIdx;
+public class NoticeVO {
+    private int noticeIdx;
 	private String noticeName;
 	private String noticeDetail;
+	private int memberIdx;
+	private String memberNick;
 	private Date noticeRegistered;
-	private int noticeView;
+	private int noticeViews;
 	private int noticeReplies;
 	private MultipartFile attach;
