+++ src/main/resources/mybatis/mapper/noticeFile-mapper.xml
@@ -29,8 +29,11 @@
 		)
 	</insert>
 	
-	<select id="get" parameterType="int" resultType="MemberProfileDto">
+	<select id="get" parameterType="int" resultType="NoticeFileDto">
 		select * from notice_file where notice_file_idx = #{notice_file_idx}
+	</select>
+	
+	<select id="list">
 	</select>
