+++ src/main/webapp/WEB-INF/views/notice/write.jsp
@@ -2,7 +2,7 @@
     pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <!DOCTYPE html>
-<form method="post">
+<form method="post" enctype="multipart/form-data">
 <table border="0">
 	<tbody>
 		<tr>
